# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Trading Alert Engine API
  description: |
    Real-time trading alert system using Alpaca Markets data.

    This API provides:
    - Alert management (view, filter, mark as read)
    - Trading rule configuration via YAML
    - Watchlist management
    - Real-time market data
    - WebSocket streaming for live updates
  version: 1.0.0
  contact:
    name: Trading Engine Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.trading-engine.local/api/v1
    description: Production server
...
paths:
...
  /rules:
    get:
      tags:
        - Rules
      summary: List all rules
      description: Retrieve all trading rules
      operationId: listRules
      responses:
        '200':
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'

    post:
      tags:
        - Rules
      summary: Create a new rule
      description: Create a new trading rule with YAML configuration
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
      responses:
        '201':
          description: Created rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '400':
          description: Invalid rule configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Rule with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rules/{rule_id}:
    get:
      tags:
        - Rules
      summary: Get rule by ID
      description: Retrieve a single rule by its ID
      operationId: getRule
      parameters:
        - name: rule_id
          in: path
          required: true
          description: The rule ID
          schema:
            type: integer
      responses:
        '200':
          description: Rule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Rules
      summary: Update rule
      description: Update an existing trading rule
      operationId: updateRule
      parameters:
        - name: rule_id
          in: path
          required: true
          description: The rule ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleUpdate'
      responses:
        '200':
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '400':
          description: Invalid rule configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Rules
      summary: Delete rule
      description: Delete a trading rule
      operationId: deleteRule
      parameters:
        - name: rule_id
          in: path
          required: true
          description: The rule ID
          schema:
            type: integer
      responses:
        '204':
          description: Rule deleted successfully
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rules/{rule_id}/toggle:
    post:
      tags:
        - Rules
      summary: Toggle rule active status
      description: Enable or disable a trading rule
      operationId: toggleRule
      parameters:
        - name: rule_id
          in: path
          required: true
          description: The rule ID
          schema:
            type: integer
      responses:
        '200':
          description: Updated rule with toggled status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /watchlist:
    get:
      tags:
        - Watchlist
      summary: Get watchlist
      description: Retrieve all symbols in the watchlist
      operationId: getWatchlist
      responses:
        '200':
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatchlistItem'

    post:
      tags:
        - Watchlist
      summary: Add to watchlist
      description: Add a symbol to the watchlist
      operationId: addToWatchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistCreate'
      responses:
        '201':
          description: Added to watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItem'
        '400':
          description: Invalid symbol
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Symbol already in watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /watchlist/{symbol}:
    delete:
      tags:
        - Watchlist
      summary: Remove from watchlist
      description: Remove a symbol from the watchlist
      operationId: removeFromWatchlist
      parameters:
        - name: symbol
          in: path
          required: true
          description: The ticker symbol
          schema:
            type: string
      responses:
        '204':
          description: Removed from watchlist
        '404':
          description: Symbol not in watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Rule:
      type: object
      required:
        - id
        - name
        - rule_type
        - config_yaml
        - is_active
        - priority
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: breakout_setup
        description:
          type: string
          nullable: true
          example: Price breakout above resistance with volume confirmation
        rule_type:
          type: string
          enum:
            - price
            - volume
            - gap
            - technical
          example: price
        config_yaml:
          type: string
          description: YAML configuration for the rule
          example: |
            conditions:
              - field: price
                operator: ">"
                value: resistance_level
        is_active:
          type: boolean
          example: true
        priority:
          type: integer
          description: Higher priority rules are evaluated first
          example: 10
        alerts_triggered:
          type: integer
          description: Number of alerts this rule has triggered
          example: 45
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RuleCreate:
      type: object
      required:
        - name
        - rule_type
        - config_yaml
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: my_custom_rule
        description:
          type: string
          maxLength: 500
        rule_type:
          type: string
          enum:
            - price
            - volume
            - gap
            - technical
        config_yaml:
          type: string
          description: YAML configuration for the rule
        is_active:
          type: boolean
          default: true
        priority:
          type: integer
          default: 0
    RuleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        config_yaml:
          type: string
        is_active:
          type: boolean
        priority:
          type: integer
    WatchlistItem:
      type: object
      required:
        - id
        - symbol
        - added_at
        - is_active
      properties:
        id:
          type: integer
          example: 1
        symbol:
          type: string
          example: AAPL
        added_at:
          type: string
          format: date-time
        is_active:
          type: boolean
          example: true
        notes:
          type: string
          nullable: true
          example: Watch for earnings breakout
    WatchlistCreate:
      type: object
      required:
        - symbol
      properties:
        symbol:
          type: string
          minLength: 1
          maxLength: 10
          example: AAPL
        notes:
          type: string
          maxLength: 500
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Alert with ID 123 not found
        details:
          type: object
          additionalProperties: true

    # WebSocket Message Schemas (for documentation purposes)
...